Warning fix (32 bits?)
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@507 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ copyfile.c
@@ -64,7 +64,7 @@ int copyfile(Stream_t *Source, Stream_t *Target)
 			else
 				fprintf(stderr,
 					"Short write %lu instead of %d\n",
-					retw, ret);
+					(unsigned long) retw, ret);
 			if(errno == ENOSPC)
 				got_signal = 1;
 			return ret;
